responsible-ai-model-detail/workbench/requirements/requirement.txt
uvicorn
pyyaml
fastapi==0.99.1
watchfiles
requests==2.25
python-multipart==0.0.5
email-validator
pymongo
python-dotenv
tensorflow
joblib
scikit-learn==1.3.0
xgboost==2.1.0
statsmodels==0.14.2
fastapi_csrf_protect
numpy==1.26.4
lib/aicloudlibs-0.1.0-py3-none-any.whl
# ../lib/aicloudlibs-0.1.0-py3-none-any.whl  # Un-Comment this line and comment above line to run locally
